double checking to make it a bit more robust

+++ setup.sh
@@ -1,11 +1,11 @@
-#!/bin/sh
+#!/bin/bash
 
 # ---------------------
 # Check root privileges 
 # ---------------------
 if [ "$EUID" -ne 0 ]; then
   echo "Please run as root"
-  exit
+  exit 1
 fi
 
 
@@ -23,7 +23,7 @@ if [[ $yn =~ ^[Yy]$ ]]; then
 
   if [[ $PASSWORD != $PASSWORD_CONFIRM ]]; then
     echo "Passwords do not match!"
-    exit
+    exit 1
   fi
 
   # create user
@@ -33,8 +33,8 @@ if [[ $yn =~ ^[Yy]$ ]]; then
   echo "${USERNAME}:${PASSWORD}" | chpasswd
 
   # give user sudo privileges (update visudo)
-  echo "%wheel ALL=(ALL) ALL" >> etc/sudoers
-  echo "User ${USERNAME} created and grated full permissions!"
+  echo "%wheel ALL=(ALL) ALL" >> /etc/sudoers
+  echo "User ${USERNAME} created and granted full permissions!"
 fi
 
 
@@ -57,12 +57,37 @@ fi
 read -p "Install yay? (y/n) " yn
 
 if [[ $yn =~ ^[Yy]$ ]]; then
-  git clone https://aur.archlinux.org/yay.git
-  cd yay
-  makepkg -si
-  cd ..
-  rm -rf yay
-  echo "Installed yay!"
+
+  # check if USERNAME is set
+  if [ -z "$USERNAME" ]; then
+    read -p "Enter username to build yay with: " TARGET_USER
+  else
+    TARGET_USER="$USERNAME"
+  fi
+
+  # check if user exists
+  if ! id "$TARGET_USER" >/dev/null 2>&1; then
+    echo "User '$TARGET_USER' does not exist! Aborting zsh setup."
+    exit 1
+  fi
+
+  # check if yay is already installed
+  if command -v yay >/dev/null 2>&1; then
+    echo "yay is already installed!"
+  else
+    USER_HOME=$(eval echo "~$TARGET_USER")
+  
+    sudo -u "$TARGET_USER" bash -c "
+      cd $USER_HOME &&
+      git clone https://aur.archlinux.org/yay.git &&
+      cd yay &&
+      makepkg -si --noconfirm &&
+      cd .. &&
+      rm -rf yay
+    "
+    
+    echo "Installed yay!"
+  fi
 fi
 
 
@@ -72,15 +97,43 @@ fi
 read -p "Install zsh and oh-my-zsh? (y/n) " yn
 
 if [[ $yn =~ ^[Yy]$ ]]; then
+
+  # check if USERNAME is set
+  if [ -z "$USERNAME" ]; then
+    read -p "Enter username, for who to install oh-my-zsh: " TARGET_USER
+  else
+    TARGET_USER="$USERNAME"
+  fi
+
+  # check if user exists
+  if ! id "$TARGET_USER" >/dev/null 2>&1; then
+    echo "User '$TARGET_USER' does not exist! Aborting zsh setup."
+    exit 1
+  fi
+
+  # get home directory of user
+  USER_HOME=$(eval echo "~$TARGET_USER")
+
+  # install zsh
   pacman -S zsh
-  curl -fsSL https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
 
-  # download custom .zshrc
-  wget "https://github.com/jannis2106/arch-setup/blob/main/.zshrc" -O .zshrc
+  # install oh-my-zsh as user
+  sudo -iu "$TARGET_USER" sh -c "$(wget -O- https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh)"
 
+  # download custom .zshrc as user
+  sudo -iu "$TARGET_USER" wget "https://raw.githubusercontent.com/jannis2106/arch-setup/refs/heads/main/.zshrc" -O .zshrc
+
+  # ensure correct ownership
+  chown "$TARGET_USER:$TARGET_USER" "$USER_HOME/.zshrc"  
+
+  # set zsh as default shell
+  chsh -s /bin/zsh "$TARGET_USER"
+
+  # install plugins
   pacman -S zsh-autosuggestions
+  
   echo "zsh and oh-my-zsh installed!"
 fi
 
 echo "Script Done!"
-exit
+exit 1
